requirements.txt
requests
beautifulsoup4
spacy
nltk
transformers
scikit-learn
fastapi
uvicorn
pandas
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2==2.9.6
python-dotenv==1.0.0
scikit-learn==1.3.0